Force a (temporary) re-layout of our presenting view controller during rotation
This allows us to capture a correct screenshot of its views on rotation,
so the partially obscured UI appears to update as expected. The 
presenting view controller is assumed to provide the entire backing 
context for this redraw (i.e.: it's assumed that simply rendering its
immediate view hierarchy to a screenshot will be sufficient to produce
the expected background view.